Crate partner_chains_demo_runtime

Modules

api

genesis_config_presets

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Structs

BlockHashCount

BlockLength

BlockProducerMetadataType

BlockWeights

FeeMultiplier

MaxTransactions

MaxValidators

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

TokenConversionRate

Version

Enums

BlockAuthor

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

Constants

DAYS

EXISTENTIAL_DEPOSIT

Existential deposit.

HOURS

MAXIMUM_BLOCK_LENGTH

MAXIMUM_BLOCK_WEIGHT

We allow for 2 seconds of compute with a 6 second average block time.

MAX_METADATA_URL_LENGTH

MILLISECS_PER_BLOCK

This determines the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

MINUTES

RUNTIME_API_VERSIONS

SLOT_DURATION

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WASM_BINARY_PATH

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AddressAssociations

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Aura

AuraConfig

Balance

Balance of an account.

Balances

BalancesConfig

BeneficiaryId

Block

Block type as expected by this runtime.

BlockNumber

An index to a block.

BlockParticipation

BlockProducerMetadata

BlockProductionLog

CrossChainPublic

Executive

Executive: handles dispatch to the various modules.

GovernedMap

GovernedMapConfig

Grandpa

GrandpaConfig

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Migrations

NativeTokenManagement

NativeTokenManagementConfig

Nonce

Index of a transaction in the chain.

PalletSession

PalletSessionConfig

Session

SessionCommitteeManagement

SessionCommitteeManagementConfig

SessionConfig

Sidechain

SidechainConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Sudo

SudoConfig

System

SystemConfig

TestHelperPallet

TestHelperPalletConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.