partner_chains_demo_runtime

Module opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Modules

cross_chain_app

Structs

CrossChainKey

CrossChainRuntimeAppPublic

SessionKeys

UncheckedExtrinsic

Simple blob to hold an extrinsic without committing to its format and ensure it is serialized correctly.

Constants

CROSS_CHAIN

Type Aliases

Block

Opaque block type.

BlockId

Opaque block identifier type.

Header

Opaque block header type.